requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
